requirements.txt
alabaster==0.7.12
alembic==1.3.3
amqp==5.0.6
appdirs==1.4.4
astroid==2.5.6
attrs==20.3.0
auth==0.5.3
Authlib==0.15.3
Babel==2.9.1
billiard==3.6.4.0
blinker==1.4
botocore==1.19.29
cachelib==0.1.1
celery==5.1.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Click==7.0
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
cors==1.0.1
cryptography==3.4.7
distlib==0.3.1
dnspython==1.16.0
docker==5.0.0
docopt==0.6.2
docutils==0.15.2
ecdsa==0.14.1
eventlet==0.31.0
falcon==3.0.1
filelock==3.0.12
Flask==1.1.1
Flask-Cors==3.0.10
Flask-Migrate==3.0.0
Flask-OAuthlib==0.9.6
Flask-Script==2.0.6
Flask-Script-Extras==0.1.4
Flask-SQLAlchemy==2.4.1
Foundations==2.0.7
future==0.18.2
gevent==21.1.2
greenlet==1.1.0
gunicorn==20.0.4
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.11.0
jmespath==0.10.0
jose==1.0.0
kombu==5.1.0
lazy-object-proxy==1.4.0
linecache2==1.0.0
Mako==1.1.1
Manager==2.0.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.23.1
oauthlib==2.1.0
ordereddict==1.1
packaging==20.9
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
psycopg2-pool==1.1
py==1.10.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.9.0
PyJWT==2.1.0
pylint==2.7.2
pylint-plugin-utils==0.6
pymongo==3.11.4
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
python-jose==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.1.0
rsa==4.7.2
s3transfer==0.3.3
simplejson==3.17.2
six==1.14.0
snowballstemmer==2.1.0
Sphinx==4.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.13
tldextract==3.1.0
toml==0.10.1
traceback2==1.4.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
unittest2==1.1.0
urllib3==1.26.4
vine==5.0.0
virtualenv==20.4.1
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==0.16.1
wrapt==1.11.1
WTForms==2.2.1
yarg==0.1.9
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
